Move EM algorithm into a separate class
